+++ src/app/(detail)/detail/[id]/page.tsx
@@ -1,0 +1,62 @@
+'use client';
+
+import { getAnimeDetail } from '@/action/anime';
+import Layout from '@/app/(default)/layout';
+import Tabs from '@/component/common/tab/tab-list';
+import DetailBanner from '@/component/detail/banner';
+import CastProductList from '@/component/detail/cast-production-list';
+import CommentContainer from '@/component/detail/comment-container';
+import EpisodeList from '@/component/detail/episode-list';
+import { DETAIL_TABS } from '@/constant/common';
+import { useQuery } from '@tanstack/react-query';
+import { useParams, useRouter, useSearchParams } from 'next/navigation';
+
+const DetailPage = () => {
+  const router = useRouter();
+  const { id } = useParams();
+  const searchParams = useSearchParams();
+  const tab = searchParams.get('tab') || 'episode';
+
+  const { data } = useQuery({
+    queryKey: ['animeDetail', id],
+    queryFn: () => getAnimeDetail(id as string),
+    enabled: !!id,
+  });
+
+  const handleChangeTab = (tab: string) => {
+    router.push(`/detail/${id}?tab=${tab}`);
+  };
+
+  return data ? (
+    <>
+      <DetailBanner
+        poster={data.poster}
+        title={data.title}
+        rating={data.averageRating}
+        runningTime={data.runningTime}
+        categories={data.categories}
+      />
+      <div className='py-5 px-5 md:px-8 flex flex-col gap-5'>
+        <Tabs
+          list={Object.entries(DETAIL_TABS).map(([id, value]) => ({
+            id,
+            value,
+          }))}
+          value={tab}
+          setValue={handleChangeTab}
+        />
+        <div>
+          {tab === 'episode' && <EpisodeList list={data.episodes} />}
+          {tab === 'cast' && <CastProductList list={data.castings} />}
+          {tab === 'comment' && (
+            <CommentContainer list={data.reviews} rating={data.averageRating} />
+          )}
+        </div>
+      </div>
+    </>
+  ) : (
+    <Layout>데이터가 없습니다.</Layout>
+  );
+};
+
+export default DetailPage;
