+++ src/app/(detail)/layout.tsx
@@ -1,0 +1,12 @@
+import Header from '@/component/common/header/header';
+
+const Layout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <>
+      <Header />
+      {children}
+    </>
+  );
+};
+
+export default Layout;
